Fix for loading DISCORD_BOT_TOKEN in a docker container
There was a bug where you still had to provide 'final.txt' when you ran inside of a docker container as the check for this file would error out if it didn't exists, even if DISCORD_BOT_TOKEN was provided through an environmental variable. This fix will attempt to load through the files, then use a provided environmental variable, and error out with a bit more verbose of a message if no token is provided still.